+++ emcorn/http/__init__.py
@@ -1,10 +1,6 @@
 __all__ = [
-    'HttpRequest',
-    'HttpResponse',
-    'HttpParser',
+    "HttpRequest", "HttpResponse"
 ]
-
 
 from .request import HttpRequest
 from .response import HttpResponse
-from .parser import HttpParser