+++ emcorn/http/_type.py
@@ -1,7 +1,0 @@
-import io
-
-class StringIO(io.StringIO):
-    
-    @property
-    def len(self):
-        return len(self.getvalue())