+++ changelog.md
@@ -1,2 +1,3 @@
 # v0.0.1
-1. http
+1. socket
+2. connection: keep-alive